Removed the preliminary closure type and ownership transfers
The closure type CLOSURE_PRELIMINARY was used for static initialization of
a blueprint's variables in LDMud 3.2 and before. Those closures needed
adaption to the real object after the object was finalized. Also the .user
field of every array and mapping in the blueprint's variables were corrected
in that step.

Since LDMud 3.3 variables are initialized after the object is created
and was given a uid, so there is no need for that mechanism anymore.